pyproject.toml
[tool.poetry]
name = "pythoneda-infrastructure-nix-flakes"
version = "0.0.1a3"
description = "Infrastructure layer of PythonEDA Nix Flakes"
authors = ["rydnr <github@acm-sl.org>"]
readme = "README.md"
homepage = "https://github.com/pythoneda-infrastructure/nix-flakes"
packages = [{include = "PythonEDAInfrastructureNixFlakes"}]

[tool.poetry.dependencies]
python = "^3.6"
pythoneda = "^0.0.1a7"
pythoneda-nix-flakes = "^0.0.1a2"
pythoneda-infrastructure-base = "^0.0.1a5"

[tool.poetry.dev-dependencies]
pytest = "^7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
